FIX: Check if file exists upfront.

+++ lib/file_store/local_store.rb
@@ -12,11 +12,10 @@ module FileStore
     def remove_file(url)
       return unless is_relative?(url)
       path = public_dir + url
+      return if !File.exists?(path)
       tombstone = public_dir + url.sub("/uploads/", "/tombstone/")
       FileUtils.mkdir_p(Pathname.new(tombstone).dirname)
       FileUtils.move(path, tombstone, :force => true)
-    rescue Errno::ENOENT
-      # don't care if the file isn't there
     end
 
     def has_been_uploaded?(url)
